Cast to float

+++ vlp/seq2seq_loader.py
@@ -473,10 +473,10 @@ class Preprocess4Seq2seqDecoder(Pipeline):
                 img_name = img_id + '.jpg'
                 bbox_img_name = '/'.join(img_path.split('/')[:7]) + '/' + img_name
                 with open(self.region_det_file_prefix + img_name + '_feats.pkl', 'rb') as region_feat_f, open(self.region_det_file_prefix + img_name +'_scores.pkl', 'rb') as region_cls_f, open(self.region_bbox_file, 'rb') as region_bbox_f:
-                    img = torch.from_numpy(pickle.load(region_feat_f, encoding="bytes"))
-                    cls_label = torch.from_numpy(pickle.load(region_cls_f, encoding="bytes"))
+                    img = torch.from_numpy(pickle.load(region_feat_f, encoding="bytes")).float()
+                    cls_label = torch.from_numpy(pickle.load(region_cls_f, encoding="bytes")).float()
                     # vis_pe = torch.cat((torch.from_numpy(pickle.load(region_bbox_f, encoding="bytes")[bbox_img_name]), cls_label.max(dim=1)[0].unsqueeze(1).double()), 1)
-                    vis_pe = torch.from_numpy(pickle.load(region_bbox_f, encoding="bytes")[bbox_img_name])
+                    vis_pe = torch.from_numpy(pickle.load(region_bbox_f, encoding="bytes")[bbox_img_name]).float()
 
                 # with h5py.File(self.region_det_file_prefix+'_feat'+img_id[-3:] +'.h5', 'r') as region_feat_f, \
                 #         h5py.File(self.region_det_file_prefix+'_cls'+img_id[-3:] +'.h5', 'r') as region_cls_f, \
@@ -502,6 +502,6 @@ class Preprocess4Seq2seqDecoder(Pipeline):
             vis_pe = torch.cat((vis_pe[:, :4], rel_area.view(-1, 1), vis_pe[:, 5:]), -1) # confident score
             normalized_coord = F.normalize(vis_pe.data[:, :5]-0.5, dim=-1)
             vis_pe = torch.cat((F.layer_norm(vis_pe, [6]), \
-                F.layer_norm(cls_label.double(), [1601])), dim=-1) # 1601 hard coded...
+                F.layer_norm(cls_label, [1601])), dim=-1) # 1601 hard coded...
 
         return (input_ids, segment_ids, position_ids, input_mask, self.task_idx, img, vis_pe)
